Get closer to a working ALSA implementation.

+++ firmware/emu/sound/sound.h
@@ -18,8 +18,8 @@ struct SoundDriver
 	void (*CloseFn)( void * object );
 	int (*SoundStateFn)( void * object );
 	SoundCBType callback;
-	int channelsPlay;
-	int channelsRec;
+	short channelsPlay;
+	short channelsRec;
 	int sps;
 	void * opaque;
