+++ firmware/emu/sound/sound_win.c
@@ -20,8 +20,8 @@ struct SoundDriverWin
 	void (*CloseFn)( void * object );
 	int (*SoundStateFn)( void * object );
 	SoundCBType callback;
-	int channelsPlay;
-	int channelsRec;
+	short channelsPlay;
+	short channelsRec;
 	int sps;
 	void * opaque;
